review: refactor out getStatsUri to URLHelper
