+++ src/main/java/uk/ac/ebi/eva/utils/URLHelper.java
@@ -21,6 +21,10 @@ import java.nio.file.Paths;
 
 public class URLHelper {
 
+    private static final String VARIANT_STATS_SUFFIX = ".variants.stats.json.gz";
+
+    private static final String SOURCE_STATS_SUFFIX = ".source.stats.json.gz";
+
     public static URI createUri(String input) throws URISyntaxException {
         URI sourceUri = new URI(input);
         if (sourceUri.getScheme() == null || sourceUri.getScheme().isEmpty()) {
@@ -29,4 +33,18 @@ public class URLHelper {
         return sourceUri;
     }
 
+    public static URI getVariantsStatsUri(String outputDirStatistics, String studyId, String fileId) throws URISyntaxException {
+        return URLHelper.createUri(
+                getStatsBaseUri(outputDirStatistics, studyId, fileId).getPath() + VARIANT_STATS_SUFFIX);
+    }
+
+    public static URI getSourceStatsUri(String outputDirStatistics, String studyId, String fileId) throws URISyntaxException {
+        return URLHelper.createUri(
+                getStatsBaseUri(outputDirStatistics, studyId, fileId).getPath() + SOURCE_STATS_SUFFIX);
+    }
+
+    public static URI getStatsBaseUri(String outputDirStatistics, String studyId, String fileId) throws URISyntaxException {
+        URI outdirUri = URLHelper.createUri(outputDirStatistics);
+        return outdirUri.resolve(MongoDBHelper.buildStorageFileId(studyId, fileId));
+    }
 }
